Feature: added axis and changed move callback
Axis property added. In tMove method we only need to prevent scrolling only if the X axis used. If we are using Y axis then we have to do the move calculations and decide after if we have to prevent the scroll or not. This is why I have to extend the default move callback with return value. This return value can indicate in this case if we have to prevent scrolling or not.

Example:

Imagine a slider with vertical animation. It changes when there is a vertical touch in progress, but when we on the last slide, we can release the "prevent scroll" to let the user scroll below the slider.